Add fuzzer test for liquidity pool entries

+++ services/horizon/internal/db2/history/liquidity_pools_batch_insert_builder_test.go
@@ -63,9 +63,9 @@ func TestAddLiquidityPool(t *testing.T) {
 		lp := lps[0]
 		cp := lPoolEntry.Body.ConstantProduct
 		tt.Assert.Equal(lPoolEntry.LiquidityPoolId, lp.PoolID)
-		tt.Assert.Equal(uint32(cp.Params.Fee), lp.Fee)
-		tt.Assert.Equal(uint64(cp.TotalPoolShares), lp.ShareCount)
-		tt.Assert.Equal(uint64(cp.PoolSharesTrustLineCount), lp.TrustlineCount)
+		tt.Assert.Equal(cp.Params.Fee, lp.Fee)
+		tt.Assert.Equal(cp.TotalPoolShares, lp.ShareCount)
+		tt.Assert.Equal(cp.PoolSharesTrustLineCount, lp.TrustlineCount)
 		tt.Assert.Len(lp.AssetReserves, 2)
 		tt.Assert.Equal(buildLiquidityPoolAssetReserves(*lPoolEntry.Body.ConstantProduct), lp.AssetReserves)
 		tt.Assert.Equal(null.NewString("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML", true), lp.Sponsor)
